Fix appointment href and change how patients are queried
The link to the identified patient's appointment was modified
based on the patient search. This was done by adding
'?id=patent_id' to the href. Resulted in an error if different
patients were searched on one page. Updated to check for this.

Also, the API included a way to include query parameters
so using that instead of doing on the server. The added benefit
is the API query matches names that are not exact, but close.